created the general component for button
